+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## next
 
 - Emulate click (single tap) and double click (double tap) in touch devices to support backdrop and image zooming (#210).
+- Ignore pointer events when not the primary button was pressed (#221).
 
 ## 1.3.0 (Oct 25, 2018)